@@ -1,49 +1,45 @@
 {
-  "name": "pwa",
+  "name": "stonemor-survey-admin",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/date-fns": "^1.3.6",
-    "@date-io/moment": "^1.3.7",
-    "@material-ui/core": "^4.0.0",
-    "@material-ui/icons": "^4.0.0",
-    "@material-ui/lab": "^4.0.0-alpha.61",
-    "@material-ui/pickers": "^3.0.0",
-    "apexcharts": "^3.35.5",
-    "apollo-boost": "^0.4.0",
-    "apollo-link-state": "^0.4.2",
-    "aws-amplify": "^1.1.28",
-    "aws-amplify-react": "^2.3.8",
-    "aws-appsync": "^1.8.0",
-    "aws-sdk": "^2.494.0",
-    "axios": "^1.1.3",
-    "chart.js": "^3.9.1",
-    "date-fns": "^2.0.0-alpha.27",
-    "graphql": "^14.3.1",
-    "graphql-tag": "^2.10.1",
+    "@apollo/client": "^3.7.1",
+    "@aws-amplify/ui-react": "^4.1.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fontsource/roboto": "^4.5.8",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.10.16",
+    "@mui/x-date-pickers": "^5.0.10",
+    "apexcharts": "^3.36.3",
+    "aws-amplify": "^5.0.4",
+    "aws-appsync-auth-link": "^3.0.7",
+    "aws-appsync-subscription-link": "^3.1.2",
+    "axios": "^1.2.1",
+    "jspdf": "^2.5.1",
     "moment": "^2.29.4",
     "qrcode.react": "^3.1.0",
-    "react": "^16.8.6",
-    "react-apollo": "^2.5.6",
-    "react-big-calendar": "^0.21.0",
-    "react-chartjs-2": "^4.3.1",
+    "react": "^18.2.0",
+    "react-apexcharts": "^1.4.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^16.8.6",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "3.0.1",
-    "uuid": "^3.3.2",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.4",
+    "react-scripts": "^5.0.1",
     "validator": "^13.7.0",
-    "victory": "^36.6.8"
+    "xlsx": "https://cdn.sheetjs.com/xlsx-0.19.1/xlsx-0.19.1.tgz"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
+    "build-run": "react-scripts build && npx serve build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
