@@ -1,34 +1,29 @@
 /* eslint-disable */
 // WARNING: DO NOT EDIT. This file is automatically generated by AWS Amplify. It will be overwritten.
 
-const awsmobile = {
-    "aws_project_region": "us-west-2",
-    "aws_cognito_identity_pool_id": "us-west-2:18d1ecae-4e3e-4db0-b5c6-fcafb88471b9",
-    "aws_cognito_region": "us-west-2",
-    "aws_user_pools_id": "us-west-2_nV8QHCQhc",
-    "aws_user_pools_web_client_id": "7d0pg869rovur5e7s48snrvj06",
-    "oauth": {},
-    "aws_cognito_username_attributes": [],
-    "aws_cognito_social_providers": [],
-    "aws_cognito_signup_attributes": [
-        "EMAIL"
-    ],
-    "aws_cognito_mfa_configuration": "OFF",
-    "aws_cognito_mfa_types": [
-        "SMS"
-    ],
-    "aws_cognito_password_protection_settings": {
-        "passwordPolicyMinLength": 8,
-        "passwordPolicyCharacters": []
-    },
-    "aws_cognito_verification_mechanisms": [
-        "EMAIL"
-    ],
-    "aws_appsync_graphqlEndpoint": "https://bskidyyyabhdxb5huxo6fhv6u4.appsync-api.us-west-2.amazonaws.com/graphql",
-    "aws_appsync_region": "us-west-2",
-    "aws_appsync_authenticationType": "API_KEY",
-    "aws_appsync_apiKey": "da2-bgfmfmwl2vajxed5iyxdfv3gnu"
+const AppSyncConfig = {
+  aws_project_region: "us-west-2",
+  aws_cognito_identity_pool_id:
+    "us-west-2:18d1ecae-4e3e-4db0-b5c6-fcafb88471b9",
+  aws_cognito_region: "us-west-2",
+  aws_user_pools_id: "us-west-2_nV8QHCQhc",
+  aws_user_pools_web_client_id: "7d0pg869rovur5e7s48snrvj06",
+  oauth: {},
+  aws_cognito_username_attributes: [],
+  aws_cognito_social_providers: [],
+  aws_cognito_signup_attributes: ["EMAIL"],
+  aws_cognito_mfa_configuration: "OFF",
+  aws_cognito_mfa_types: ["SMS"],
+  aws_cognito_password_protection_settings: {
+    passwordPolicyMinLength: 8,
+    passwordPolicyCharacters: [],
+  },
+  aws_cognito_verification_mechanisms: ["EMAIL"],
+  aws_appsync_graphqlEndpoint:
+    "https://bskidyyyabhdxb5huxo6fhv6u4.appsync-api.us-west-2.amazonaws.com/graphql",
+  aws_appsync_region: "us-west-2",
+  aws_appsync_authenticationType: "API_KEY",
+  aws_appsync_apiKey: "da2-bgfmfmwl2vajxed5iyxdfv3gnu",
 };
 
-
-export default awsmobile;
+export default AppSyncConfig;
